requirements.txt
pycryptodome
pytest
pyautogui
